General rework
(individual commits stashed unfortunately)

This reworks the interface between high and low level layers to
make pipelining easier and optionally allow for a fifo to be
inserted (though none is yet).

There's also an optional delay on sampling the input data (doing
it on negedge) which helps with some setups with routing delays.

Signed-off-by: Benjamin Herrenschmidt <benh@kernel.crashing.org>